pyproject.toml
[project]
name = "pyaigress"
version = "0.1.0"
description = ""
authors = [
    {name="Taokhan"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "arq (>=0.27.0,<0.28.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "redis (==5.0)"
]

[tool.poetry]
packages = [{include = "pyaigress", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
